+++ src/test/java/org/stellar/sdk/TransactionTest.java
@@ -38,7 +38,6 @@ import org.stellar.sdk.xdr.LedgerEntryType;
 import org.stellar.sdk.xdr.LedgerFootprint;
 import org.stellar.sdk.xdr.LedgerKey;
 import org.stellar.sdk.xdr.SCVal;
-import org.stellar.sdk.xdr.SignerKey;
 import org.stellar.sdk.xdr.SorobanResources;
 import org.stellar.sdk.xdr.SorobanTransactionData;
 import org.stellar.sdk.xdr.Uint256;